+++ HISTORY.md
@@ -1,5 +1,14 @@
 # Revision History for `matdb`
 
+## Revision 0.0.3
+
+- Debugged the modulation sub-config generation.
+- Debugged the phonon amplitude calibration.
+- Added status messages for database DFT execution progress.
+- Debugged job array submission.
+- Added several new statements to prevent re-building of setup and cleanup phases in
+  databases that are already finished.
+
 ## Revision 0.0.2
 
 - Added support for job file submission.